SRxのPATHを通しながらrunするように実装、confファイルの生成場所の修正、rm機能の引数数チェックの実装

+++ rm_project.sh
@@ -3,9 +3,14 @@
 #
 #
 
+if [ $# -ne 1 ]
+then
+	exit 1
+fi
+
 PR_NAME=$1
 
-CON_LIST=(`docker ps --filter "name=pr_$PR_NAME" --format "{{.Names}}"`)
+CON_LIST=(`docker ps -a --filter "name=pr_$PR_NAME" --format "{{.Names}}"`)
 for con in ${CON_LIST[@]}
 do
 	docker kill $con
